Create EXAMPLE OF DIFFERENCE
Well There Would Not Even Be These Issues 
If Minecraft And Forge Had Been Build On 
Lua Instead Of Java Trust Me Im A Senior 
Java Developer And I Have Noted That Java 
Has Massive Holes In It And As I Am A Lua 
Coder Too I Know Lua Has Les Holes In It  
But I See This Is Not A Java Hole Issue It 
Is Just That Mine&Blade Is Poorly Coded And 
Utilizes In Different Way Than Mekanism. 
Ps. By The Way More Mods Would Be 
Incompatible If Minecraft And Forge Had Been 
Coded In Lua But If There Would Then Be 
A Special Edition Of Lua Named 
EXAMPLE-NAME-HERE Or Something 
Then It Would Not Be As Big Of 
A Mess In Lua But As Minecraft 
And Forge Are Java Based
And Incompatible With Lua Itd 
Be Impossible To Do A Port Now 
And If EXAMPLE-NAME-HERE Would 
Be A Java Code Type Then Itd 
Be Easier To Get Mods 
Compatible Now As It 
Would Not Be A Mess Of 
A Different Utilizations Of 
Java Scripts And Such